simplify kokoro's implementation
