+++ s2s_pipeline.py
@@ -34,10 +34,7 @@ from arguments_classes.melo_tts_arguments import MeloTTSHandlerArguments
 from arguments_classes.open_api_language_model_arguments import OpenApiLanguageModelHandlerArguments
 from arguments_classes.facebookmms_tts_arguments import FacebookMMSTTSHandlerArguments
 from arguments_classes.pocket_tts_arguments import PocketTTSHandlerArguments
-from arguments_classes.kokoro_tts_arguments import (
-    KokoroTTSHandlerArguments,
-    KokoroMLXTTSHandlerArguments,
-)
+from arguments_classes.kokoro_tts_arguments import KokoroTTSHandlerArguments
 import torch
 import nltk
 from rich.console import Console
@@ -104,7 +101,6 @@ def parse_arguments():
             FacebookMMSTTSHandlerArguments,
             PocketTTSHandlerArguments,
             KokoroTTSHandlerArguments,
-            KokoroMLXTTSHandlerArguments,
         )
     )
 
@@ -141,7 +137,7 @@ def optimal_mac_settings(mac_optimal_settings: Optional[str], *handler_kwargs):
             if hasattr(kwargs, "llm"):
                 kwargs.llm = "mlx-lm"
             if hasattr(kwargs, "tts"):
-                kwargs.tts = "kokoro-mlx"
+                kwargs.tts = "kokoro"
 
 def check_mac_settings(module_kwargs):
     if platform == "darwin":
@@ -197,7 +193,6 @@ def prepare_all_args(
     facebook_mms_tts_handler_kwargs,
     pocket_tts_handler_kwargs,
     kokoro_tts_handler_kwargs,
-    kokoro_mlx_tts_handler_kwargs,
 ):
     prepare_module_args(
         module_kwargs,
@@ -215,7 +210,6 @@ def prepare_all_args(
         facebook_mms_tts_handler_kwargs,
         pocket_tts_handler_kwargs,
         kokoro_tts_handler_kwargs,
-        kokoro_mlx_tts_handler_kwargs,
     )
 
     rename_args(whisper_stt_handler_kwargs, "stt")
@@ -232,7 +226,6 @@ def prepare_all_args(
     rename_args(facebook_mms_tts_handler_kwargs, "facebook_mms")
     rename_args(pocket_tts_handler_kwargs, "pocket_tts")
     rename_args(kokoro_tts_handler_kwargs, "kokoro")
-    rename_args(kokoro_mlx_tts_handler_kwargs, "kokoro_mlx")
 
 
 def initialize_queues_and_events():
@@ -266,7 +259,6 @@ def build_pipeline(
     facebook_mms_tts_handler_kwargs,
     pocket_tts_handler_kwargs,
     kokoro_tts_handler_kwargs,
-    kokoro_mlx_tts_handler_kwargs,
     queues_and_events,
 ):
     stop_event = queues_and_events["stop_event"]
@@ -332,7 +324,7 @@ def build_pipeline(
 
     stt = get_stt_handler(module_kwargs, stop_event, spoken_prompt_queue, text_prompt_queue, whisper_stt_handler_kwargs, faster_whisper_stt_handler_kwargs, paraformer_stt_handler_kwargs, mlx_audio_whisper_stt_handler_kwargs, parakeet_tdt_stt_handler_kwargs)
     lm = get_llm_handler(module_kwargs, stop_event, text_prompt_queue, lm_response_queue, language_model_handler_kwargs, open_api_language_model_handler_kwargs, mlx_language_model_handler_kwargs)
-    tts = get_tts_handler(module_kwargs, stop_event, lm_response_queue, send_audio_chunks_queue, should_listen, parler_tts_handler_kwargs, melo_tts_handler_kwargs, chat_tts_handler_kwargs, facebook_mms_tts_handler_kwargs, pocket_tts_handler_kwargs, kokoro_tts_handler_kwargs, kokoro_mlx_tts_handler_kwargs)
+    tts = get_tts_handler(module_kwargs, stop_event, lm_response_queue, send_audio_chunks_queue, should_listen, parler_tts_handler_kwargs, melo_tts_handler_kwargs, chat_tts_handler_kwargs, facebook_mms_tts_handler_kwargs, pocket_tts_handler_kwargs, kokoro_tts_handler_kwargs)
 
     return ThreadManager([*comms_handlers, vad, stt, lm, tts])
 
@@ -447,7 +439,7 @@ def get_llm_handler(
         raise ValueError("The LLM should be either transformers or mlx-lm")
 
 
-def get_tts_handler(module_kwargs, stop_event, lm_response_queue, send_audio_chunks_queue, should_listen, parler_tts_handler_kwargs, melo_tts_handler_kwargs, chat_tts_handler_kwargs, facebook_mms_tts_handler_kwargs, pocket_tts_handler_kwargs, kokoro_tts_handler_kwargs, kokoro_mlx_tts_handler_kwargs):
+def get_tts_handler(module_kwargs, stop_event, lm_response_queue, send_audio_chunks_queue, should_listen, parler_tts_handler_kwargs, melo_tts_handler_kwargs, chat_tts_handler_kwargs, facebook_mms_tts_handler_kwargs, pocket_tts_handler_kwargs, kokoro_tts_handler_kwargs):
     if module_kwargs.tts == "parler":
         from TTS.parler_handler import ParlerTTSHandler
         return ParlerTTSHandler(
@@ -512,17 +504,8 @@ def get_tts_handler(module_kwargs, stop_event, lm_response_queue, send_audio_chu
             setup_args=(should_listen,),
             setup_kwargs=vars(kokoro_tts_handler_kwargs),
         )
-    elif module_kwargs.tts == "kokoro-mlx":
-        from TTS.kokoro_mlx_handler import KokoroMLXTTSHandler
-        return KokoroMLXTTSHandler(
-            stop_event,
-            queue_in=lm_response_queue,
-            queue_out=send_audio_chunks_queue,
-            setup_args=(should_listen,),
-            setup_kwargs=vars(kokoro_mlx_tts_handler_kwargs),
-        )
     else:
-        raise ValueError("The TTS should be either parler, melo, chatTTS, facebookMMS, pocket, kokoro, or kokoro-mlx")
+        raise ValueError("The TTS should be either parler, melo, chatTTS, facebookMMS, pocket, or kokoro")
 
 
 def main():
@@ -545,7 +528,6 @@ def main():
         facebook_mms_tts_handler_kwargs,
         pocket_tts_handler_kwargs,
         kokoro_tts_handler_kwargs,
-        kokoro_mlx_tts_handler_kwargs,
     ) = parse_arguments()
 
     setup_logger(module_kwargs.log_level)
@@ -566,7 +548,6 @@ def main():
         facebook_mms_tts_handler_kwargs,
         pocket_tts_handler_kwargs,
         kokoro_tts_handler_kwargs,
-        kokoro_mlx_tts_handler_kwargs,
     )
 
     queues_and_events = initialize_queues_and_events()
@@ -590,7 +571,6 @@ def main():
         facebook_mms_tts_handler_kwargs,
         pocket_tts_handler_kwargs,
         kokoro_tts_handler_kwargs,
-        kokoro_mlx_tts_handler_kwargs,
         queues_and_events,
     )
